Refactor API scope
