+++ Common/Sources/DataServices/UserService.swift
@@ -1,0 +1,12 @@
+import Combine
+import Foundation
+
+/// A protocol that defines the API for interacting with the user data.
+public protocol UserService {
+    associatedtype SomeUser: User
+    
+    /// Retrieves a user by their ID.
+    /// - Parameter userID: The ID of the user.
+    /// - Returns: A `StreamDataOperation` that emits the user data and continues to observe for future changes.
+    func getUser(byID userID: String) -> StreamDataOperation<SomeUser>
+}
